+++ tasks/omp/tsatsyn_a_radix_sort_simple_merge/func_tests/main.cpp
@@ -7,7 +7,7 @@
 #include <vector>
 
 #include "core/task/include/task.hpp"
-#include "omp/tsatsyn_a_radix_sort_simple_merge_omp/include/ops_omp.hpp"
+#include "omp/tsatsyn_a_radix_sort_simple_merge/include/ops_omp.hpp"
 
 namespace {
 std::vector<double> GetRandomVector(int sz, int a, int b) {
