+++ tasks/omp/tsatsyn_a_radix_sort_simple_merge/perf_tests/main.cpp
@@ -9,7 +9,7 @@
 
 #include "core/perf/include/perf.hpp"
 #include "core/task/include/task.hpp"
-#include "omp/tsatsyn_a_radix_sort_simple_merge_omp/include/ops_omp.hpp"
+#include "omp/tsatsyn_a_radix_sort_simple_merge/include/ops_omp.hpp"
 
 namespace {
 std::vector<double> GetRandomVector(int sz, int a, int b) {
